feat: change props extend BucketProps, update test
